Add standalone chat interface
